crispv3/requirements_jay.txt
altair==5.0.1
attrs==24.2.0
backports.zoneinfo==0.2.1
blinker==1.6.3
cachetools==5.5.2
charset-normalizer==3.4.1
click==8.0.4
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
fonttools==4.38.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==0.5.1
google-cloud==0.34.0
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos>=1.56.2,<2.0.0
idna==3.10
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.17.3
kiwisolver==1.4.5
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
monai==1.1.0
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plotly==5.7.0
proto-plus==1.26.1
protobuf==3.20.2
#pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b1
Pygments==2.17.2
Pympler==1.1
pyparsing==3.1.4
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2025.1
pytz-deprecation-shim==0.1.0.post0
#PyWavelets==1.3.0
# pywin32==308
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
semver==3.0.4
six==1.17.0
smmap==5.0.2
streamlit==1.23.1
tenacity==8.2.3
threadpoolctl==3.1.0
tifffile==2021.11.2
toml==0.10.2
toolz==0.12.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
treelib==1.6.1
typing_extensions==4.7.1
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.0.7
validators==0.20.0
watchdog==3.0.0
wincertstore==0.2
zipp==3.15.0
albumentations==1.3.1
